Bash file to pull out evaluation metrics from automatic frequency post-processing
Run this bash file in the folder where the output from RunAllOntologiesConceptMapper.java is located (make this file executable and then run it). 

The bash file aggregates all the evaluation metrics for each concept frequency threshold run over the range for each ontology and outputs a file for each ontology with this information with the name of the ontology followed by the extension "_info.txt". 

The output of this will be used in the getMetrics_Optimization.py file to find the optimal concept frequency threshold that maximizes precision, recall, or F-measure.